# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Workbench API
  version: 0.1.0
paths:
  /health/live:
    get:
      operationId: getLiveness
      responses:
        "200":
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /v1/tasks:
    get:
      operationId: listTasks
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/TaskStatus"
      responses:
        "200":
          description: Owner-scoped task list
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/tasks/{id}:
    patch:
      operationId: updateTask
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Updated Task with incremented version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/captures:
    post:
      operationId: createCapture
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaptureRequest"
      responses:
        "201":
          description: Capture was created or an identical idempotent request was reused
          headers:
            X-Request-ID:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capture"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: The idempotency key was already used for different input
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/inbox:
    get:
      operationId: listInbox
      responses:
        "200":
          description: Owner-scoped pending Candidate list
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Candidate"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/candidates/{id}/reject:
    post:
      operationId: rejectCandidate
      parameters:
        - $ref: "#/components/parameters/CandidateId"
      responses:
        "200":
          description: Candidate was rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateMutationResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/candidates/{id}/accept:
    post:
      operationId: acceptCandidate
      parameters:
        - $ref: "#/components/parameters/CandidateId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptCandidateRequest"
      responses:
        "201":
          description: Candidate was accepted and a Task was created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptCandidateResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    CandidateId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  headers:
    RequestId:
      description: Request identifier used to correlate API logs
      schema:
        type: string
  responses:
    Unauthorized:
      description: Authentication is required
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: The request parameters are invalid
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The owner-scoped resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The resource state changed and the command was not applied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: An unexpected server error occurred
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
          const: ok
    TaskStatus:
      type: string
      enum: [backlog, in_progress, blocked, done]
    CaptureStatus:
      type: string
      enum: [queued, processing, processed, failed]
    CandidateStatus:
      type: string
      enum: [pending_review, accepted, rejected]
    CreateCaptureRequest:
      type: object
      required: [idempotencyKey, inputText, sourceType]
      properties:
        idempotencyKey:
          type: string
          minLength: 1
          maxLength: 120
        inputText:
          type: string
          minLength: 1
          maxLength: 250
        sourceType:
          type: string
          minLength: 1
          maxLength: 40
        externalRef:
          type: [string, "null"]
          maxLength: 200
        sourceUrl:
          type: [string, "null"]
          maxLength: 500
        excerpt:
          type: [string, "null"]
          maxLength: 250
        consentScope:
          type: [string, "null"]
          maxLength: 80
    Capture:
      type: object
      required: [id, status, requestId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/CaptureStatus"
        requestId:
          type: string
        createdAt:
          type: string
          format: date-time
    Candidate:
      type: object
      required: [id, captureId, proposedTitle, labels, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        captureId:
          type: string
          format: uuid
        proposedTitle:
          type: string
          minLength: 1
          maxLength: 200
        proposedDescription:
          type: [string, "null"]
          maxLength: 10000
        proposedProjectId:
          type: [string, "null"]
          format: uuid
        labels:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 40
        status:
          $ref: "#/components/schemas/CandidateStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CandidateMutationResult:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/CandidateStatus"
    AcceptCandidateRequest:
      type: object
      additionalProperties: false
      required: [title, labels]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: [string, "null"]
          maxLength: 10000
        projectId:
          type: [string, "null"]
          format: uuid
        labels:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 40
    AcceptCandidateResult:
      type: object
      required: [candidateId, taskId, status]
      properties:
        candidateId:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        status:
          type: string
          const: accepted
    Task:
      type: object
      required: [id, title, status, priority, labels, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: [string, "null"]
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 10000
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          type: string
          enum: [none, low, medium, high]
        labels:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 40
        dueAt:
          type: [string, "null"]
          format: date-time
        archivedAt:
          type: [string, "null"]
          format: date-time
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateTaskRequest:
      type: object
      additionalProperties: false
      required: [version, title, description, status, priority, labels]
      properties:
        version:
          type: integer
          minimum: 1
        projectId:
          type: [string, "null"]
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 10000
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          type: string
          enum: [none, low, medium, high]
        labels:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 40
        dueAt:
          type: [string, "null"]
          format: date-time
    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string